MPI+X backends work, tests pass with gnu
